requirements.txt
numpy>=1.21.0
torch>=1.12.0
gymnasium>=0.28.0
matplotlib>=3.5.0
pyyaml>=6.0
pandas>=1.3.0

